Swap path font color from black to cyan
Black was impossible to see in my Emacs terminal, it shows up as more of
a gray in iTerm, but the Emacs vterm is darker. I think that's just a
Doom setting that I should probably play with (because I want it to look
the same as my iTerm session), but this looks better anyways so I'll
just change the theme.

Signed-off-by: Michael Leonard <m-leonard@ti.com>